requirements.txt
face-recognition>=1.3.0
numpy>=1.24
Pillow>=10.0
tqdm>=4.60
rawpy>=0.17
imageio>=2.37.0
